+++ src/shared/sdoc-commons/services/sdoc-solr.adapter.ts
@@ -19,7 +19,7 @@ export class SDocSolrAdapter extends GenericSolrAdapter<SDocRecord, SDocSearchFo
             'rate_pers_motive_i', 'rate_pers_schwierigkeit_i', 'rate_pers_wichtigkeit_i',
             'rate_tech_overall_s', 'rate_tech_ks_s', 'rate_tech_firn_s', 'rate_tech_gletscher_s', 'rate_tech_klettern_s',
             'rate_tech_bergtour_s', 'rate_tech_schneeschuh_s',
-            'gpstracks_basefile_s', 'keywords_txt', 'loc_lochirarchie_s', 'loc_lochirarchie_ids_s', 'name_s', 'type_s',
+            'gpstracks_basefile_s', 'keywords_txt', 'loc_lochirarchie_s', 'loc_lochirarchie_ids_s', 'name_s', 'type_s', 'persons_txt',
             'actiontype_ss', 'subtype_s', 'i_fav_url_txt', 'v_fav_url_txt'],
         facetConfigs: {
             'actiontype_ss': {
@@ -54,6 +54,8 @@ export class SDocSolrAdapter extends GenericSolrAdapter<SDocRecord, SDocSearchFo
                 'f.month_is.facet.sort': 'index'
             },
             'persons_txt': {
+                'f.persons_txt.limit': '-1',
+                'f.persons_txt.facet.sort': 'index'
             },
             'playlists_txt': {
             },